+++ dev/e2e_app/integration_test/permissions/clear_permissions_test.dart
@@ -1,3 +1,5 @@
+import 'package:e2e_app/keys.dart';
+
 import '../common.dart';
 
 const _timeout = Duration(seconds: 5); // to avoid timeouts on CI
@@ -11,7 +13,9 @@ void main() {
     await _requestAndGrantCameraPermission($);
     await _requestAndGrantMicrophonePermission($);
     await _requestAndGrantLocationPermission($);
+    await _requestAndGrantGalleryPermission($);
   });
+
   patrol('grants various permissions 2', ($) async {
     await createApp($);
 
@@ -20,12 +24,13 @@ void main() {
     await _requestAndGrantCameraPermission($);
     await _requestAndGrantMicrophonePermission($);
     await _requestAndGrantLocationPermission($);
+    await _requestAndGrantGalleryPermission($);
   });
 }
 
 Future<void> _requestAndGrantCameraPermission(PatrolIntegrationTester $) async {
-  expect($(#camera).$(#statusText).text, 'Not granted');
-  await $('Request camera permission').tap();
+  expect($(K.cameraPermissionTile).$(K.statusText).text, 'Not granted');
+  await $(K.requestCameraPermissionButton).tap();
   if (await $.native.isPermissionDialogVisible(timeout: _timeout)) {
     await $.native.grantPermissionWhenInUse();
     await $.pump();
@@ -35,8 +40,8 @@ Future<void> _requestAndGrantCameraPermission(PatrolIntegrationTester $) async {
 Future<void> _requestAndGrantMicrophonePermission(
   PatrolIntegrationTester $,
 ) async {
-  expect($(#microphone).$(#statusText).text, 'Not granted');
-  await $('Request microphone permission').tap();
+  expect($(K.microphonePermissionTile).$(K.statusText).text, 'Not granted');
+  await $(K.requestMicrophonePermissionButton).tap();
   if (await $.native.isPermissionDialogVisible(timeout: _timeout)) {
     await $.native.grantPermissionOnlyThisTime();
     await $.pump();
@@ -46,8 +51,19 @@ Future<void> _requestAndGrantMicrophonePermission(
 Future<void> _requestAndGrantLocationPermission(
   PatrolIntegrationTester $,
 ) async {
-  expect($(#location).$(#statusText).text, 'Not granted');
-  await $('Request location permission').tap();
+  expect($(K.locationPermissionTile).$(K.statusText).text, 'Not granted');
+  await $(K.requestLocationPermissionButton).tap();
+  if (await $.native.isPermissionDialogVisible(timeout: _timeout)) {
+    await $.native.grantPermissionOnlyThisTime();
+    await $.pump();
+  }
+}
+
+Future<void> _requestAndGrantGalleryPermission(
+  PatrolIntegrationTester $,
+) async {
+  expect($(K.galleryPermissionTile).$(K.statusText).text, 'Not granted');
+  await $(K.requestGalleryPermissionButton).tap();
   if (await $.native.isPermissionDialogVisible(timeout: _timeout)) {
     await $.native.grantPermissionOnlyThisTime();
     await $.pump();