+++ packages/patrol_cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased 
+
+- Add support for de, fr and pl locales for native methods that operates on strings (#2659)
+- Add support for gallery permission dialog on iOS 17 (#2659)
+
 ## 3.9.0
 
 - Re-land: Do not pass flavor as dart-define when building. (#2636)